make message (part) size quint64 instead of (u)int
Message and message part sizes were treated in the code as uints
in general, but sometimes as an int (e.g., via toInt). As per
RFC 3501, we should at least make sure the uint (32 bit unsigned
integers) range is covered.

Some servers are capabale of serving messages larger than 32 bit
integers, i.e., allow 64 bit sizes. To future-proof Trojita
against this, we use quint64. (Note that currently sizes in
Trojita are limited practically by the use of QByteArray, which
has int size.) This implies updating the SQL schema version to v7.

This patch does not explicitly limit attachment sizes to 32 bit
(unsigned) integers. (Qt file sizes are usually expressed as
64 bit signed integers.)

+++ src/Imap/Model/Cache.h
@@ -52,7 +52,7 @@ public:
         /** @short The IMAP's INTERNALDATE */
         QDateTime internalDate;
         /** @short RFC822.SIZE */
-        uint size;
+        quint64 size;
         /** @short Serialized form of BODYSTRUCTURE
 
         Due to the complex nature of BODYSTRUCTURE and the way we use, simly
